@@ -9,37 +9,42 @@
     "sitemap": "node sitemap-generator"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/vue-fontawesome": "^2.0.2",
-    "bulma": "^0.9.2",
-    "cheerio": "^1.0.0-rc.6",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/vue-fontawesome": "^3.0.0-5",
+    "bulma": "^0.9.3",
+    "cheerio": "^1.0.0-rc.10",
+    "core-js": "^3.19.1",
     "register-service-worker": "^1.7.2",
-    "vue": "^2.6.12",
-    "vue-analytics": "^5.22.1",
-    "vue-class-component": "^7.2.6",
-    "vue-meta": "^2.4.0",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.5.1",
+    "vue": "^3.2.21",
+    "vue-gtag-next": "^1.14.0",
+    "vue-meta": "^3.0.0-alpha.7",
+    "vue-router": "^4.0.12",
     "vue-scrollto": "^2.20.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.12",
-    "@vue/cli-plugin-pwa": "^4.5.12",
-    "@vue/cli-plugin-typescript": "^4.5.12",
-    "@vue/cli-service": "^4.5.12",
-    "css-loader": "^5.2.4",
-    "file-loader": "^6.2.0",
-    "image-webpack-loader": "^7.0.1",
-    "node-sass": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.3.1",
+    "@typescript-eslint/parser": "^5.3.1",
+    "@vue/cli-plugin-babel": "~4.5.15",
+    "@vue/cli-plugin-eslint": "~4.5.15",
+    "@vue/cli-plugin-pwa": "~4.5.15",
+    "@vue/cli-plugin-router": "~4.5.15",
+    "@vue/cli-plugin-typescript": "~4.5.15",
+    "@vue/cli-service": "~4.5.15",
+    "@vue/compiler-sfc": "^3.2.21",
+    "@vue/eslint-config-prettier": "^6.0.0",
+    "@vue/eslint-config-typescript": "^9.0.1",
+    "@vue/server-renderer": "^3.2.21",
+    "eslint": "^8.2.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-vue": "^8.0.3",
     "prerender-spa-plugin": "^3.4.0",
-    "sass-loader": "^10.1.1",
+    "prettier": "^2.4.1",
+    "sass": "^1.43.4",
+    "sass-loader": "^10.2.0",
     "sitemap-generator": "^8.5.1",
-    "typescript": "^4.2.4",
-    "url-loader": "^4.1.1",
-    "vue-style-loader": "^4.1.3",
-    "vue-svg-loader": "^0.16.0",
-    "vue-template-compiler": "^2.6.12"
+    "typescript": "^4.4.4",
+    "vue-loader": "^16.8.3"
   }
 }
