@@ -1,22 +1,22 @@
-const SitemapGenerator = require('sitemap-generator');
-const _sitemapUrl = 'https://www.vueseofriendlyspatemplate.com';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const SitemapGenerator = require("sitemap-generator");
+const SITEMAP_URL = "https://www.vueseofriendlyspatemplate.com/";
 
-// create generator
-const generator = SitemapGenerator(_sitemapUrl, {
+const generator = SitemapGenerator(SITEMAP_URL, {
   lastMod: true,
   stripQuerystring: true,
   filepath: `${__dirname}/public/sitemap.xml`
 });
 
-// register event listener (SUCCESS => sitemaps created)
-generator.on('done', () => {
-  console.log(`sitemap.xml successfully created for URL: ${_sitemapUrl}\n`);
+// Register event listener for 'done'
+generator.on("done", () => {
+  console.log(`sitemap.xml successfully created for URL: ${SITEMAP_URL}\n`);
 });
 
-// register event listener (ERROR => { code: 404, message: 'Not found.', url: 'http://example.com/foo' })
-generator.on('error', (error) => {
-  console.error(`${JSON.stringify(error)}\n`);
+// Register event listener for 'error'
+// ErrorMessage has shape => { code: 404, message: 'Not found.', url: 'http://example.com/foo' }
+generator.on("error", err => {
+  console.error(`${JSON.stringify(err)}\n`);
 });
 
-// start the crawler
 generator.start();
