@@ -1,4 +1,4 @@
-﻿<template functional>
+﻿<template>
   <footer class="footer">
     <div class="buttons">
       <a
@@ -10,22 +10,27 @@
       >
         <font-awesome-icon :icon="['fab', 'github']" />
       </a>
-      <a class="button is-medium" aria-label="Twitter">
+      <a
+        target="_blank"
+        aria-label="Twitter"
+        class="button is-medium"
+        rel="noopener noreferrer"
+        href="https://github.com/based-ghost"
+      >
         <font-awesome-icon :icon="['fab', 'twitter']" />
       </a>
-      <a class="button is-medium" aria-label="Medium">
+      <a
+        target="_blank"
+        aria-label="Medium"
+        class="button is-medium"
+        rel="noopener noreferrer"
+        href="https://github.com/based-ghost"
+      >
         <font-awesome-icon :icon="['fab', 'medium-m']" />
       </a>
     </div>
-    <div class="content">
-      VueSeoFriendlySpaTemplate.com &copy; 2019
+    <div class="content has-text-centered">
+      Copyright © 2021 vueseofriendlyspa.com
     </div>
   </footer>
 </template>
-
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-
-@Component
-export default class Footer extends Vue {}
-</script>
