@@ -1,10 +1,4 @@
-import Navbar from './Navbar.vue';
-import AppFooter from './Footer.vue';
-import BackToTop from './BackToTop.vue';
-
-
-export {
-  Navbar,
-  AppFooter,
-  BackToTop
-};
+export { default as Alert } from './Alert.vue';
+export { default as Navbar } from './Navbar.vue';
+export { default as AppFooter } from './Footer.vue';
+export { default as BackToTop } from './BackToTop.vue';
