@@ -1,32 +1,25 @@
-export type FeatureInfo = Readonly<{
+export type Feature = Readonly<{
   description: string;
-  packageName?: string;
+  package_name?: string;
 }>;
 
-export const FeatureInfoConfig: { [key: string]: FeatureInfo } = {
-  typescript: {
-    description: 'TypeScript + class-based component syntax'
+export const FEATURES: Feature[] = [
+  {
+    description: 'UI styled with Bulma + SASS + Font Awesome 5 (svg-core)'
   },
-  css: {
-    description: 'UI styled with Bulma CSS + SASS + Font Awesome 5 (svg-core)'
-  },
-  pwa: {
+  {
     description: 'Configured as a (PWA) Progressive Web App'
   },
-  vmeta: {
+  {
     description: 'Meta tags dynamically handled per route using',
-    packageName: 'vue-meta'
+    package_name: 'vue-meta'
   },
-  vanalytics: {
-    description: 'Google Analytics ready to go and easily configurable using',
-    packageName: 'vue-analytics'
+  {
+    description: 'Global Site Tag plugin already configured',
+    package_name: 'vue-gtag-next'
   },
-  prerender: {
+  {
     description: 'Configured to serve prerendered html using',
-    packageName: 'prerender-spa-plugin'
-  },
-  vsvg: {
-    description: 'Configured to allow loading of your own SVG images using',
-    packageName: 'vue-svg-loader'
-  },
-};
+    package_name: 'prerender-spa-plugin'
+  }
+];
